merge.c adapted for floats

+++ Problems/Merge/merge.c
@@ -5,14 +5,14 @@
 // Merges two subarrays of arr[].
 // First subarray is arr[l..m]
 // Second subarray is arr[m+1..r]
-void merge(int arr[], int l, int m, int r)
+void merge(float arr[], int l, int m, int r)
 {
 	int i, j, k;
 	int n1 = m - l + 1;
 	int n2 = r - m;
 
 	// Create temp arrays
-	int L[n1], R[n2];
+	float L[n1], R[n2];
 
 	// Copy data to temp arrays L[] and R[]
 	for (i = 0; i < n1; i++)
@@ -55,7 +55,7 @@ void merge(int arr[], int l, int m, int r)
 
 // l is for left index and r is right index of the
 // sub-array of arr to be sorted
-void mergeSort(int arr[], int l, int r)
+void mergeSort(float arr[], int l, int r)
 {
 	if (l < r) {
 		int m = l + (r - l) / 2;
@@ -69,18 +69,18 @@ void mergeSort(int arr[], int l, int r)
 }
 
 // Function to print an array
-void printArray(int A[], int size)
+void printArray(float A[], int size)
 {
 	int i;
 	for (i = 0; i < size; i++)
-		printf("%d ", A[i]);
+		printf("%f ", A[i]);
 	printf("\n");
 }
 
 // Driver code
 int main()
 {
-	int arr[] = { 12, 11, 13, 5, 6, 7 };
+	float arr[] = { 12.0, 11.0, 13.0, 5.0, 6.0, 7.0 };
 	int arr_size = sizeof(arr) / sizeof(arr[0]);
 
 	printf("Given array is \n");
